Use size_t where possible

+++ src/filewriter.h
@@ -6,6 +6,6 @@
 class FileWriter
 {
 public:
-    static bool writeMapToFile(std::vector<char> map, int width, int height, const std::string& filename, const std::string& solution);
-    static std::vector<char> finalizeMap(std::vector<char> map, int width, int height);
+    static bool writeMapToFile(std::vector<char> map, size_t width, size_t height, const std::string& filename, const std::string& solution);
+    static std::vector<char> finalizeMap(std::vector<char> map, size_t width, size_t height);
 };
